Lesson 3 Bean scope
First scope - singltone. This is default scope.
Bean in this scaope is always created if we indicate the context, it's common f>
Example in applicationContext2

Second scope - prototype.
Bean is created if we turn to the Spring Container with getBean method
Each circulation creates a new object
Example in applecationContext2